Move to matrices benchmarks
